requirements.txt
asgiref==3.8.1
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
beautifulsoup4==4.11.1
boto3==1.34.98
botocore==1.34.98
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
channels==3.0.1
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==3.0.2
Deprecated==1.2.13
dill==0.3.6
Django==3.2.15
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-cors-headers==3.10.1
django-environ==0.9.0
django-filter==22.1
django-storages==1.14.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
flake8==5.0.4
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
msgpack==1.0.4
packaging==21.3
phonenumbers==8.13.35
Pillow==9.3.0
platformdirs==2.5.4
priority==1.3.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycountry==23.12.11
pycparser==2.21
pyflakes==2.5.0
PyJWT==1.7.1
pylint==2.15.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2022.1
PyYAML==6.0.1
redis==4.3.4
requests==2.28.1
s3transfer==0.10.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==5.0.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.12
wget==3.2
wrapt==1.16.0
zope.interface==6.3


### python version 3.10